feat: update life cycle for qwen vl ocr
