+++ README.md
@@ -127,7 +127,7 @@ Predefined domain options include：["Technology","Finance","Health","Education"
 # Advanced PDF parsing (requires MinerU)
 dm = DataMax(file_path="complex.pdf", use_mineru=True)
 
-# PDF OCR parsing (requires API credentials)(Only Qwen-VL-OCR model is currently available)
+# PDF OCR parsing (requires API credentials)(Only Qwen-VL-OCR[-latest] model is currently available)
 dm = DataMax(
     file_path="scanned.pdf", 
     use_qwen_vl_ocr=True,
@@ -139,8 +139,8 @@ dm = DataMax(
 # Word to Markdown conversion
 dm = DataMax(file_path="document.docx", to_markdown=True)
 
-# Image OCR
-dm = DataMax(file_path="image.jpg", use_ocr=True)
+# Image OCR （need install mineru）
+dm = DataMax(file_path="image.jpg")
 ```
 ### Batch Processing
 ```python
