+++ README_zh.md
@@ -79,7 +79,8 @@ qa_data = dm.get_pre_label(
 # PDF高级解析（需要MinerU）
 dm = DataMax(file_path="complex.pdf", use_mineru=True)
 
-# PDF OCR解析（需要API凭证）（当前仅支持qwen-vl-ocr模型）
+# PDF OCR解析（需要API凭证）（当前仅支持qwen-vl-ocr[-latest]模型）
+# tips: 请注意，use_qwen_vl_ocr参数用于图片或pdf的文字识别。
 dm = DataMax(
     file_path="scanned.pdf", 
     use_qwen_vl_ocr=True,
@@ -91,8 +92,8 @@ dm = DataMax(
 # Word转Markdown
 dm = DataMax(file_path="document.docx", to_markdown=True)
 
-# 图片OCR
-dm = DataMax(file_path="image.jpg", use_mineru=True)
+# 图片 (需要安装mineru)
+dm = DataMax(file_path="image.jpg")
 ```
 
 ### 批处理解析
